+++ eRecruiter.Api/Parameters/Job/SocialMedia.cs
@@ -1,0 +1,7 @@
+﻿namespace eRecruiter.Api.Parameters
+{
+    public class SocialMedia
+    {
+        public string Description { get; set; }
+    }
+}