+++ eRecruiter.Api/Responses/Job/JobCreationResponse.cs
@@ -1,0 +1,9 @@
+﻿using System.Net;
+
+namespace eRecruiter.Api.Responses
+{
+    public class JobCreationResponse
+    {
+        public int Id { get; set; }
+    }
+}